refactor: InputParser 패키지 위치 변경 및 메소드 Static 전환
