+++ src/test/java/racingcar/InputParserTest.java
@@ -5,22 +5,20 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.engine.InputParser;
+import racingcar.engine.view.InputParser;
 
 public class InputParserTest {
 
-    private final InputParser inputParser = new InputParser();
-
     @Test
     public void 입력값_파싱() {
-        String[] carNames = inputParser.splitByComma("pobi,jaws,jayde");
+        String[] carNames = InputParser.splitByComma("pobi,jaws,jayde");
         Assertions.assertThat(carNames).containsExactly("pobi", "jaws", "jayde");
     }
 
     @ParameterizedTest
     @CsvSource(value = {"1:1", "12:12", "500:500"}, delimiter = ':')
     public void 실행횟수_파싱(String input, int expected) {
-        int answer = inputParser.parseStringToPositiveInt(input);
+        int answer = InputParser.parseStringToPositiveInt(input);
         Assertions.assertThat(answer).isEqualTo(expected);
     }
 
@@ -28,7 +26,7 @@ public class InputParserTest {
     @ValueSource(strings = {"asd", "", "   "})
     public void 실행횟수_포맷_검증(String input) {
         Assertions.assertThatExceptionOfType(RuntimeException.class).isThrownBy(() -> {
-            inputParser.parseStringToPositiveInt(input);
+            InputParser.parseStringToPositiveInt(input);
         }).withMessage("1 이상의 숫자만 입력해야합니다.");
     }
 
@@ -36,7 +34,7 @@ public class InputParserTest {
     @ValueSource(strings = {"0", "-12", "-2321312"})
     public void 실행횟수_양수_검증(String input) {
         Assertions.assertThatExceptionOfType(RuntimeException.class).isThrownBy(() -> {
-            inputParser.parseStringToPositiveInt(input);
+            InputParser.parseStringToPositiveInt(input);
         }).withMessage("음수는 입력할 수 없습니다.");
     }
 }
